Keep the code consistent.

+++ toxcore/tox.c
@@ -265,18 +265,18 @@ void tox_set_sends_receipts(void *tox, int friendnumber, int yesno)
 /* Return the number of friends in the instance m.
  * You should use this to determine how much memory to allocate
  * for copy_friendlist. */
-size_t tox_count_friendlist(void *tox)
+uint32_t tox_count_friendlist(void *tox)
 {
     Messenger *m = tox;
     return count_friendlist(m);
 }
 
 /* Copy a list of valid friend IDs into the array out_list.
- * If out_list is NULL, returns -1.
+ * If out_list is NULL, returns 0.
  * Otherwise, returns the number of elements copied.
  * If the array was too small, the contents
  * of out_list will be truncated to list_size. */
-size_t tox_copy_friendlist(void *tox, int *out_list, size_t list_size)
+uint32_t tox_copy_friendlist(void *tox, int *out_list, uint32_t list_size)
 {
     Messenger *m = tox;
     return copy_friendlist(m, out_list, list_size);
