+++ toxcore/tox.h
@@ -283,14 +283,14 @@ void tox_set_sends_receipts(Tox *tox, int friendnumber, int yesno);
 /* Return the number of friends in the instance m.
  * You should use this to determine how much memory to allocate
  * for copy_friendlist. */
-size_t tox_count_friendlist(void *tox);
-    
+uint32_t tox_count_friendlist(void *tox);
+
 /* Copy a list of valid friend IDs into the array out_list.
- * If out_list is NULL, returns -1.
+ * If out_list is NULL, returns 0.
  * Otherwise, returns the number of elements copied.
  * If the array was too small, the contents
  * of out_list will be truncated to list_size. */
-size_t tox_copy_friendlist(void *tox, int *out_list, size_t list_size);
+uint32_t tox_copy_friendlist(void *tox, int *out_list, uint32_t list_size);
 
 /* Set the function that will be executed when a friend request is received.
  *  Function format is function(uint8_t * public_key, uint8_t * data, uint16_t length)
